Move the rest of the lookup helpers to activityLogHelpers.js
